Pass exceptions in log context as specified in PSR-3
PSR-3 specifies[0] that when logging an exception, the exception object can
be passed in the 'exception' context key so that handlers can access it.
This improves the utility of handlers that read exceptions during processing.
My motivation for making this change is that `RavenHandler`'s back-end parses
stacktraces to group together errors with different messages but the same
execution path.

This change doesn't provide any utility to text based logs, as the exception
object renders as a string:

    ... "exception":"[object] (ErrorException: Undefined variable: hello at framework/view/SSViewer.php:158)" ...

[0] https://github.com/php-fig/fig-standards/blob/master/accepted/PSR-3-logger-interface.md#13-context