Fix unit tests not running, fix unit tests on PHP 5.5
It was not possible to run the unit tests as a mock of PHPUnit_Test_Case
in the SilverStripe framework folder was overriding the actual PHPUnit
class.

PHP 5.5 also appears to be stricter about the parameters for ErrorException;
a fatal error is thrown if the filename or line number parameters are falsey.